Conflicts:
	spec/tags/ruby/core/encoding/converter/primitive_convert_tags.txt
	spec/tags/ruby/core/encoding/converter/primitive_errinfo_tags.txt

+++ spec/ruby/optional/capi/ext/enumerator_spec.c
@@ -9,7 +9,7 @@ extern "C" {
 VALUE enumerator_spec_rb_enumeratorize(int argc, VALUE *argv, VALUE self) {
   VALUE obj, meth, args;
   rb_scan_args(argc, argv, "2*", &obj, &meth, &args);
-  return rb_enumeratorize(obj, meth, RARRAY_LEN(args), RARRAY_PTR(args));
+  return rb_enumeratorize(obj, meth, (int)RARRAY_LEN(args), RARRAY_PTR(args));
 }
 #endif
